move nearest backward
