+++ crates/cubek-interpolate/src/components/global/mod.rs
@@ -1,3 +1,5 @@
 mod base;
+mod nearest_backward;
 
 pub use base::*;
+pub use nearest_backward::*;
