+++ crates/cubek-interpolate/src/components/global/nearest_backward.rs
@@ -1,17 +1,14 @@
-use super::super::{linear_layout, shape_divmod};
-use crate::{InterpolateError, definition::NearestMode};
-use cubecl::{calculate_cube_count_elemwise, prelude::*};
+use crate::definition::NearestMode;
+use cubecl::prelude::*;
 use cubecl::{
+    ir::StorageType,
     num_traits::Zero,
-    std::{
-        FastDivmod,
-        tensor::layout::{linear::LinearLayout, *},
-    },
-    tensor_vector_size_parallel,
+    std::FastDivmod,
+    std::tensor::layout::{linear::LinearLayout, *},
 };
 
 #[cube(launch_unchecked, address_type = "dynamic")]
-fn interpolate_nearest_backward_kernel<F: Float, N: Size>(
+pub fn execute_interpolate_nearest_backward<F: Float, N: Size>(
     grad: &Tensor<Vector<F, N>>,
     output: &mut Tensor<Vector<F, N>>,
     shape_out: Sequence<FastDivmod<usize>>,
@@ -87,47 +84,4 @@ fn end_index<F: Float>(
     #[comptime] nearest_mode: NearestMode,
 ) -> usize {
     start_index::<F>(input_index + 1, output_size, input_size, nearest_mode)
-}
-
-pub(crate) fn interpolate_nearest_backward_launch<R: Runtime>(
-    client: &ComputeClient<R>,
-    out_grad: TensorBinding<R>,
-    output: TensorBinding<R>,
-    nearest_mode: NearestMode,
-    dtype: StorageType,
-) -> Result<(), InterpolateError> {
-    let vector_size = tensor_vector_size_parallel(
-        client.io_optimized_vector_sizes(dtype.size()),
-        &out_grad.shape,
-        &out_grad.strides,
-        out_grad.shape.len() - 1,
-    );
-    let shape_out = shape_divmod(&output);
-    let out_layout = linear_layout(&output, vector_size);
-
-    let working_units = output.shape.iter().product::<usize>() / vector_size as usize;
-    let cube_dim = CubeDim::new(client, working_units);
-    let cube_count = calculate_cube_count_elemwise(client, working_units, cube_dim);
-
-    let address_type = out_grad
-        .required_address_type(dtype.size())
-        .max(output.required_address_type(dtype.size()));
-
-    unsafe {
-        interpolate_nearest_backward_kernel::launch_unchecked(
-            client,
-            cube_count,
-            cube_dim,
-            address_type,
-            vector_size,
-            out_grad.into_tensor_arg(),
-            output.clone().into_tensor_arg(),
-            shape_out,
-            out_layout,
-            nearest_mode,
-            dtype,
-        )
-    };
-
-    Ok(())
 }
