--- crates/cubek-interpolate/src/kernel/backward/mod.rs
@@ -1,3 +1,0 @@
-mod nearest;
-
-pub(crate) use nearest::interpolate_nearest_backward_launch;
