+++ crates/cubek-interpolate/src/launch/base.rs
@@ -1,15 +1,24 @@
 use crate::{
     InterpolateError,
     {
-        components::global::execute_interpolate,
-        definition::{InterpolateForwardProblem, InterpolateOptions, accumulator_dtype},
+        components::global::{execute_interpolate, execute_interpolate_nearest_backward},
+        definition::{
+            InterpolateForwardProblem, InterpolateOptions, NearestMode, accumulator_dtype,
+        },
         launch::InterpolateStrategy,
         routines::{
             ForwardRoutine, GlobalMemoryRoutine, InterpolateBlueprint, SharedMemoryRoutine,
         },
     },
 };
-use cubecl::{prelude::*, std::FastDivmod, tensor_vector_size_parallel};
+use cubecl::{
+    calculate_cube_count_elemwise,
+    ir::{ElemType, StorageType, UIntKind},
+    prelude::*,
+    std::FastDivmod,
+    std::tensor::layout::linear::{LinearLayoutLaunch, LinearViewLayoutLaunch},
+    tensor_vector_size_parallel,
+};
 
 pub fn interpolate_launch<R: Runtime>(
     client: &ComputeClient<R>,
@@ -106,4 +115,68 @@ fn get_cube_shape<R: Runtime>(
     cube_shape.push(num_vectors);
     cube_shape.push(cubes_per_batch);
     cube_shape
+}
+
+pub fn interpolate_nearest_backward_launch<R: Runtime>(
+    client: &ComputeClient<R>,
+    out_grad: TensorBinding<R>,
+    output: TensorBinding<R>,
+    nearest_mode: NearestMode,
+    dtype: StorageType,
+) -> Result<(), InterpolateError> {
+    let vector_size = tensor_vector_size_parallel(
+        client.io_optimized_vector_sizes(dtype.size()),
+        &out_grad.shape,
+        &out_grad.strides,
+        out_grad.shape.len() - 1,
+    );
+    let shape_out = shape_divmod(&output);
+    let out_layout = linear_layout(&output, vector_size);
+
+    let working_units = output.shape.iter().product::<usize>() / vector_size as usize;
+    let cube_dim = CubeDim::new(client, working_units);
+    let cube_count = calculate_cube_count_elemwise(client, working_units, cube_dim);
+
+    let address_type = out_grad
+        .required_address_type(dtype.size())
+        .max(output.required_address_type(dtype.size()));
+
+    unsafe {
+        execute_interpolate_nearest_backward::launch_unchecked(
+            client,
+            cube_count,
+            cube_dim,
+            address_type,
+            vector_size,
+            out_grad.into_tensor_arg(),
+            output.clone().into_tensor_arg(),
+            shape_out,
+            out_layout,
+            nearest_mode,
+            dtype,
+        )
+    };
+
+    Ok(())
+}
+
+fn shape_divmod<R: Runtime>(binding: &TensorBinding<R>) -> SequenceArg<R, FastDivmod<usize>> {
+    let mut out_seq = SequenceArg::new();
+    for dim in binding.shape.iter() {
+        out_seq.push(*dim);
+    }
+    out_seq
+}
+
+fn linear_layout<R: Runtime>(
+    binding: &TensorBinding<R>,
+    vector_size: usize,
+) -> LinearLayoutLaunch<R> {
+    LinearLayoutLaunch::from_shape_strides(
+        binding.shape.clone(),
+        binding.strides.clone(),
+        // Don't care about type size, only vector size.
+        Type::new(StorageType::Scalar(ElemType::UInt(UIntKind::U32))).with_vector_size(vector_size),
+        LinearViewLayoutLaunch::new(),
+    )
 }
